I think this is now python package-able
