+++ pyproject.toml
@@ -1,0 +1,22 @@
+[build-system]
+requires = ["setuptools>=61.0"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "testirequesti"
+version = "0.0.1"
+authors = [
+  { name="Uri", email="Uri.Cohen.Waisel@Gmail.com" },
+]
+description = "Mock requests for testing"
+readme = "README.md"
+requires-python = ">=3.8"
+classifiers = [
+    "Programming Language :: Python :: 3",
+    "License :: GNU General Public License v3.0",
+    "Operating System :: OS Independent",
+]
+
+[project.urls]
+Homepage = "https://github.com/UriCW/TestiPy"
+Issues = "https://github.com/UriCW/TestiPy/issues"
