+++ requirements.txt
@@ -1,18 +1,12 @@
-atomicwrites==1.3.0
-attrs==19.1.0
-certifi==2019.6.16
-chardet==3.0.4
-idna==2.8
-importlib-metadata==0.19
-more-itertools==7.2.0
-packaging==19.1
-pathlib2==2.3.4
-pluggy==0.12.0
-py==1.10.0
-pyparsing==2.4.2
-pytest==5.1.1
-requests==2.22.0
-six==1.12.0
-urllib3==1.26.5
-wcwidth==0.1.7
-zipp==0.6.0
+build==1.2.1
+certifi==2024.2.2
+charset-normalizer==3.3.2
+idna==3.7
+iniconfig==2.0.0
+packaging==24.0
+pluggy==1.4.0
+pyproject_hooks==1.0.0
+pytest==8.1.1
+requests==2.31.0
+testirequesti @ file:///home/uri/Development/TestiPy
+urllib3==2.2.1
